updated scripts to work with multiple users

+++ link_conf.sh
@@ -1,4 +1,21 @@
-ln -s ~/.dotfiles/.zshrc ~/
-ln -s ~/.dotfiles/.vimrc ~/
-ln -s ~/.dotfiles/.tmux.conf ~/
-ln -s ~/.dotfiles/starship.toml ~/.config/
+#!/bin/bash
+
+if [ -e ~/.zshrc ]
+then
+    ln -s ~/.dotfiles/.zshrc ~/
+fi
+
+if [ -e ~/.vimrc ]
+then
+    ln -s ~/.dotfiles/.vimrc ~/
+fi
+
+if [ -e ~/.tmux.conf ]
+then
+    ln -s ~/.dotfiles/.tmux.conf ~/
+fi
+
+if [ -e ~/.config/starship.toml ]
+then
+    ln -s ~/.dotfiles/starship.toml ~/.config/
+fi
